Refactored the testing files, put testing modules in own subdirectory, added a testing file for each group of functions
