+++ test/unit_testing/dashboard.test.js
@@ -1,0 +1,1 @@
+var module = require('../../app/js/js_functions/dashboard_functions.js');
