+++ test/unit_testing/googlesignin.test.js
@@ -1,0 +1,1 @@
+var module = require('../../app/js/js_functions/googlesignin_functions.js');
