+++ test/unit_testing/signin.test.js
@@ -1,0 +1,1 @@
+var module = require('../../app/js/js_functions/signin_functions.js');
