+++ test/unit_testing/signup.test.js
@@ -1,0 +1,1 @@
+var module = require('../../app/js/js_functions/signup_functions.js');
